Pet Bullet mode
Pet이 1초마다 Bullet 을 쏠 수 있도록 설정

+++ Source/NierAutomata/Public/HJMainCharacter.h
@@ -48,6 +48,6 @@ public:
 	float u;
 	float r;
 
-	int speed = 800;
+	float speed = 800;
 
 };
